@@ -13,42 +13,70 @@
   },
   "homepage": "https://github.com/plouc/mozaik-ext-github",
   "main": "./src/components/index.js",
+  "keywords": [
+    "github",
+    "mozaik",
+    "widget",
+    "extension",
+    "dashboard"
+  ],
   "engines": {
     "node": "0.12.x"
   },
   "dependencies": {
-    "babelify": "^6.1.3",
+    "babelify": "7.2.0",
+    "babel-core": "6.7.2",
+    "babel-preset-es2015": "6.6.0",
+    "babel-preset-react": "6.5.0",
     "chalk": "^1.1.0",
     "convict": "^0.6.1",
-    "lodash": "^3.2.0",
+    "lodash": "4.6.1",
     "moment": "^2.10.3",
     "moment-duration-format": "^1.3.0",
-    "react": "^0.13.3",
-    "react-mixin": "^1.7.0",
-    "reflux": "^0.2.5",
+    "react-mixin": "3.0.4",
+    "reflux": "0.4.0",
     "superagent": "^1.2.0",
     "superagent-bluebird-promise": "^0.6.0"
   },
   "devDependencies": {
-    "babel": "^5.8.12",
-    "babel-eslint": "^1.0.6",
-    "chai": "^3.2.0",
-    "eslint": "^0.15.1",
-    "jsdom": "^2.0.0",
-    "mocha": "^2.2.5",
-    "mockery": "^1.4.0",
-    "sinon": "^1.15.4"
+    "ava": "0.13.0",
+    "babel-eslint": "5.0.0",
+    "babel-register": "6.7.2",
+    "coveralls": "2.11.8",
+    "enzyme": "2.1.0",
+    "eslint": "2.2.0",
+    "eslint-plugin-react": "4.2.2",
+    "mockery": "1.4.1",
+    "nyc": "6.1.1",
+    "react": "^0.13.3",
+    "sinon": "1.17.3"
   },
   "peerDependencies": {
-    "mozaik": "1.x"
+    "mozaik": ">=1.0.11",
+    "react": "^0.13.3"
   },
   "scripts": {
-    "test": "mocha --opts mocha.opts ./src/tests/**/*.test.js",
-    "eslint": "eslint ./src/*"
+    "eslint": "eslint --ext .js --ext .jsx ./src/*",
+    "test": "ava",
+    "test-cover": "nyc ava",
+    "coveralls": "nyc report --reporter=text-lcov | coveralls",
+    "cover-report": "nyc report --reporter=lcov && open coverage/lcov-report/index.html"
   },
   "browserify": {
     "transform": [
       "babelify"
+    ]
+  },
+  "ava": {
+    "babel": "inherit",
+    "require": [
+      "babel-register"
+    ]
+  },
+  "nyc": {
+    "extension": [
+      ".js",
+      ".jsx"
     ]
   }
 }
