@@ -1,53 +1,46 @@
-import React           from 'react';
-import Reflux          from 'reflux';
-import Mozaik          from 'mozaik/browser';
-import StatusIcon      from './StatusIcon.jsx';
-import StatusTimestamp from './StatusTimestamp.jsx';
+import React, { Component, PropTypes } from 'react';
+import reactMixin                      from 'react-mixin';
+import { ListenerMixin }               from 'reflux';
+import Mozaik                          from 'mozaik/browser';
+import StatusIcon                      from './StatusIcon.jsx';
+import StatusTimestamp                 from './StatusTimestamp.jsx';
 
 
-export default React.createClass({
-    displayName: 'Status',
+class Status extends Component {
+    constructor(props) {
+        super(props);
 
-    mixins: [
-        Reflux.ListenerMixin,
-        Mozaik.Mixin.ApiConsumer
-    ],
-
-    getInitialState() {
-        return {
-            status: null
-        };
-    },
+        this.state = { status: null };
+    }
 
     getApiRequest() {
-        return {
-            id: 'github.status'
-        };
-    },
+        return { id: 'github.status' };
+    }
 
     onApiData(status) {
-        this.setState({
-            status: status
-        });
-    },
+        this.setState({ status });
+    }
 
     render() {
-        var widgetBodyNode = (<div className="widget__body" />);
+        const { status } = this.state;
 
-        if (this.state.status) {
-            var messageNode = null;
+        let content = null;
+        if (status) {
+            let messageNode = null;
 
-            if (this.state.status.status !== 'good') {
-                messageNode = (<span className="github__status__current__message">{ this.state.status.body }</span>);
+            if (status.status !== 'good') {
+                messageNode = (
+                    <span className="github__status__current__message">
+                        {status.body}
+                    </span>
+                );
             }
 
-            widgetBodyNode = (
-                <div className="widget__body">
-                    <div className="github__status__current">
-                        <StatusIcon status={ this.state.status.status } message={ this.state.status.body } />
-                        { messageNode }
-                        <StatusTimestamp timestamp={ this.state.status.created_on } />
-                    </div>
+            content = (
+                <div className="github__status__current">
+                    <StatusIcon status={status.status} message={status.body} />
+                    {messageNode}
+                    <StatusTimestamp timestamp={status.created_on} />
                 </div>
             );
         }
@@ -55,11 +48,21 @@ export default React.createClass({
         return (
             <div>
                 <div className="widget__header">
-                    Github <span className="widget__header__subject">Status</span>
+                    Github <span className="widget__header__subject">Stakljkjltus</span>
                     <i className="fa fa-github-alt" />
                 </div>
-                { widgetBodyNode }
+                <div className="widget__body">
+                    {content}
+                </div>
             </div>
         );
     }
-});
+}
+
+Status.displayName = 'Status';
+
+reactMixin(Status.prototype, ListenerMixin);
+reactMixin(Status.prototype, Mozaik.Mixin.ApiConsumer);
+
+
+export default Status;
