@@ -9,9 +9,8 @@ const  { Treemap }                     = Mozaik.Component;
 class IssueLabelsTreemap extends Component {
     constructor(props) {
         super(props);
-        this.state = {
-            labels: []
-        };
+
+        this.state = { labels: [] };
     }
 
     getApiRequest() {
@@ -19,29 +18,22 @@ class IssueLabelsTreemap extends Component {
 
         return {
             id:     `github.issueLabelsAggregations.${ _.pluck(labels, 'name').join('.') }`,
-            params: {
-                labels:     labels,
-                repository: repository
-            }
+            params: { repository, labels }
         };
     }
 
     onApiData(labels) {
-        this.setState({
-            labels: labels
-        });
+        this.setState({ labels });
     }
 
     render() {
-        let { labels } = this.state;
+        const { labels } = this.state;
 
-        let data = labels.map(label => {
-            return {
-                label: label.name,
-                count: label.count,
-                color: label.color
-            };
-        });
+        const data = labels.map(label => ({
+            label: label.name,
+            count: label.count,
+            color: label.color
+        }));
 
         return (
             <div>
@@ -57,8 +49,11 @@ class IssueLabelsTreemap extends Component {
     }
 }
 
+IssueLabelsTreemap.displayName = 'IssueLabelsTreemap';
+
 IssueLabelsTreemap.propTypes = {
-    labels: PropTypes.arrayOf(PropTypes.shape({
+    repository: PropTypes.string.isRequired,
+    labels:     PropTypes.arrayOf(PropTypes.shape({
         name:  PropTypes.string,
         color: PropTypes.string
     })).isRequired
@@ -67,4 +62,5 @@ IssueLabelsTreemap.propTypes = {
 reactMixin(IssueLabelsTreemap.prototype, ListenerMixin);
 reactMixin(IssueLabelsTreemap.prototype, Mozaik.Mixin.ApiConsumer);
 
-export { IssueLabelsTreemap as default };
+
+export default IssueLabelsTreemap;
