@@ -1,17 +1,33 @@
-import React from 'react';
+import React, { Component, PropTypes } from 'react';
 
-export default React.createClass({
-    displayName: 'RepositoryContributorStat',
 
+class RepositoryContributorStat extends Component {
     render() {
+        const { contributor: {author, total} } = this.props;
+
         return (
             <div className="list__item github__repository-contributors_stats__item">
-                <img src={this.props.contributor.author.avatar_url} />
-                {this.props.contributor.author.login}&nbsp;
+                <img src={author.avatar_url}/>
+                {author.login}&nbsp;
                 <span className="github__repository-contributors_stats__item__count">
-                    {this.props.contributor.total}&nbsp;<i className="fa fa-dot-circle-o" />
+                    {total}&nbsp;<i className="fa fa-dot-circle-o"/>
                 </span>
             </div>
         );
     }
-});
+}
+
+RepositoryContributorStat.displayName = 'RepositoryContributorStat';
+
+RepositoryContributorStat.propTypes = {
+    contributor: PropTypes.shape({
+        total:  PropTypes.number.isRequired,
+        author: PropTypes.shape({
+            login:      PropTypes.string.isRequired,
+            avatar_url: PropTypes.string.isRequired
+        }).isRequired
+    }).isRequired
+};
+
+
+export default RepositoryContributorStat;
