@@ -9,6 +9,7 @@ import Mozaik                          from 'mozaik/browser';
 class TopCommitter extends Component {
     constructor(props) {
         super(props);
+
         this.state = {
             topCommitter: null,
             since:        null,
@@ -17,7 +18,7 @@ class TopCommitter extends Component {
     }
 
     getApiRequest() {
-        let { repository, frequency } = this.props;
+        const { repository, frequency } = this.props;
 
         let since;
         let until;
@@ -35,11 +36,7 @@ class TopCommitter extends Component {
 
         return {
             id:     `github.repositoryCommits.${ repository }.${ since }.${ until }`,
-            params: {
-                repository: repository,
-                since:      since,
-                until:      until
-            }
+            params: { repository, since, until }
         };
     }
 
@@ -67,17 +64,15 @@ class TopCommitter extends Component {
             return;
         }
 
-        let topCommiter = _.max(committers, 'commitCount');
-        topCommiter = _.extend(topCommiter.user, { commitCount: topCommiter.commitCount });
+        let topCommitter = _.max(committers, 'commitCount');
+        topCommitter = _.extend(topCommitter.user, { commitCount: topCommitter.commitCount });
 
-        this.setState({
-            topCommitter: topCommiter
-        });
+        this.setState({ topCommitter });
     }
 
     render() {
-        let { repository, title } = this.props;
-        let { topCommitter }      = this.state;
+        const { repository, title } = this.props;
+        const { topCommitter }      = this.state;
 
         let topCommitterNode = <div className="widget__body"/>;
         if (topCommitter) {
@@ -111,9 +106,12 @@ class TopCommitter extends Component {
     }
 }
 
+TopCommitter.displayName = 'TopComitter';
+
 TopCommitter.propTypes = {
     repository: PropTypes.string.isRequired,
-    frequency:  PropTypes.oneOf(['daily', 'weekly'])
+    frequency:  PropTypes.oneOf(['daily', 'weekly']),
+    title:      PropTypes.string
 };
 
 TopCommitter.defaultProps = {
@@ -123,4 +121,5 @@ TopCommitter.defaultProps = {
 reactMixin(TopCommitter.prototype, ListenerMixin);
 reactMixin(TopCommitter.prototype, Mozaik.Mixin.ApiConsumer);
 
-export { TopCommitter as default };
+
+export default TopCommitter;
